+++ ext/libavif/ext/aom/aom_dsp/x86/intrapred_sse2.c
@@ -112,12 +112,12 @@ void aom_dc_predictor_4x8_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_4(above);
   sum_above = _mm_add_epi16(sum_left, sum_above);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 6;
   sum = divide_using_multiply_shift(sum, 2, DC_MULTIPLIER_1X2);
 
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
-  const uint32_t pred = _mm_cvtsi128_si32(row);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(row);
   dc_store_4xh(pred, 8, dst, stride);
 }
 
@@ -127,12 +127,12 @@ void aom_dc_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_4(above);
   sum_above = _mm_add_epi16(sum_left, sum_above);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 10;
   sum = divide_using_multiply_shift(sum, 2, DC_MULTIPLIER_1X4);
 
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
-  const uint32_t pred = _mm_cvtsi128_si32(row);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(row);
   dc_store_4xh(pred, 16, dst, stride);
 }
 
@@ -142,7 +142,7 @@ void aom_dc_predictor_8x4_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_8(above);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 6;
   sum = divide_using_multiply_shift(sum, 2, DC_MULTIPLIER_1X2);
 
@@ -156,10 +156,10 @@ void aom_dc_predictor_8x16_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_8(above);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 12;
   sum = divide_using_multiply_shift(sum, 3, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_8xh(&row, 16, dst, stride);
 }
 
@@ -169,10 +169,10 @@ void aom_dc_predictor_8x32_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_8(above);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 20;
   sum = divide_using_multiply_shift(sum, 3, DC_MULTIPLIER_1X4);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_8xh(&row, 32, dst, stride);
 }
 
@@ -182,10 +182,10 @@ void aom_dc_predictor_16x4_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_16_sse2(above);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 10;
   sum = divide_using_multiply_shift(sum, 2, DC_MULTIPLIER_1X4);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_16xh(&row, 4, dst, stride);
 }
 
@@ -195,10 +195,10 @@ void aom_dc_predictor_16x8_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_16_sse2(above);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 12;
   sum = divide_using_multiply_shift(sum, 3, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_16xh(&row, 8, dst, stride);
 }
 
@@ -208,10 +208,10 @@ void aom_dc_predictor_16x32_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_16_sse2(above);
   sum_above = _mm_add_epi16(sum_left, sum_above);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 24;
   sum = divide_using_multiply_shift(sum, 4, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_16xh(&row, 32, dst, stride);
 }
 
@@ -221,10 +221,10 @@ void aom_dc_predictor_16x64_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i sum_above = dc_sum_16_sse2(above);
   sum_above = _mm_add_epi16(sum_left, sum_above);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 40;
   sum = divide_using_multiply_shift(sum, 4, DC_MULTIPLIER_1X4);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_16xh(&row, 64, dst, stride);
 }
 
@@ -234,10 +234,10 @@ void aom_dc_predictor_32x8_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_8(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 20;
   sum = divide_using_multiply_shift(sum, 3, DC_MULTIPLIER_1X4);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_32xh(&row, 8, dst, stride);
 }
 
@@ -247,10 +247,10 @@ void aom_dc_predictor_32x16_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_16_sse2(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 24;
   sum = divide_using_multiply_shift(sum, 4, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_32xh(&row, 16, dst, stride);
 }
 
@@ -260,10 +260,10 @@ void aom_dc_predictor_32x64_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_64(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 48;
   sum = divide_using_multiply_shift(sum, 5, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_32xh(&row, 64, dst, stride);
 }
 
@@ -273,10 +273,10 @@ void aom_dc_predictor_64x64_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_64(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 64;
   sum /= 128;
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_64xh(&row, 64, dst, stride);
 }
 
@@ -286,10 +286,10 @@ void aom_dc_predictor_64x32_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_32_sse2(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 48;
   sum = divide_using_multiply_shift(sum, 5, DC_MULTIPLIER_1X2);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_64xh(&row, 32, dst, stride);
 }
 
@@ -299,10 +299,10 @@ void aom_dc_predictor_64x16_sse2(uint8_t *dst, ptrdiff_t stride,
   const __m128i sum_left = dc_sum_16_sse2(left);
   sum_above = _mm_add_epi16(sum_above, sum_left);
 
-  uint32_t sum = _mm_cvtsi128_si32(sum_above);
+  uint32_t sum = (uint32_t)_mm_cvtsi128_si32(sum_above);
   sum += 40;
   sum = divide_using_multiply_shift(sum, 4, DC_MULTIPLIER_1X4);
-  const __m128i row = _mm_set1_epi8((uint8_t)sum);
+  const __m128i row = _mm_set1_epi8((int8_t)sum);
   dc_store_64xh(&row, 16, dst, stride);
 }
 
@@ -319,7 +319,7 @@ void aom_dc_top_predictor_4x8_sse2(uint8_t *dst, ptrdiff_t stride,
   sum_above = _mm_shufflelo_epi16(sum_above, 0);
   sum_above = _mm_packus_epi16(sum_above, sum_above);
 
-  const uint32_t pred = _mm_cvtsi128_si32(sum_above);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(sum_above);
   dc_store_4xh(pred, 8, dst, stride);
 }
 
@@ -333,7 +333,7 @@ void aom_dc_top_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   sum_above = _mm_shufflelo_epi16(sum_above, 0);
   sum_above = _mm_packus_epi16(sum_above, sum_above);
 
-  const uint32_t pred = _mm_cvtsi128_si32(sum_above);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(sum_above);
   dc_store_4xh(pred, 16, dst, stride);
 }
 
@@ -523,7 +523,7 @@ void aom_dc_left_predictor_4x8_sse2(uint8_t *dst, ptrdiff_t stride,
   sum_left = _mm_shufflelo_epi16(sum_left, 0);
   sum_left = _mm_packus_epi16(sum_left, sum_left);
 
-  const uint32_t pred = _mm_cvtsi128_si32(sum_left);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(sum_left);
   dc_store_4xh(pred, 8, dst, stride);
 }
 
@@ -538,7 +538,7 @@ void aom_dc_left_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   sum_left = _mm_shufflelo_epi16(sum_left, 0);
   sum_left = _mm_packus_epi16(sum_left, sum_left);
 
-  const uint32_t pred = _mm_cvtsi128_si32(sum_left);
+  const uint32_t pred = (uint32_t)_mm_cvtsi128_si32(sum_left);
   dc_store_4xh(pred, 16, dst, stride);
 }
 
@@ -990,26 +990,26 @@ void aom_h_predictor_4x8_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i row1 = _mm_shufflelo_epi16(left_col, 0x55);
   __m128i row2 = _mm_shufflelo_epi16(left_col, 0xaa);
   __m128i row3 = _mm_shufflelo_epi16(left_col, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
   dst += stride;
   left_col = _mm_unpackhi_epi64(left_col, left_col);
   row0 = _mm_shufflelo_epi16(left_col, 0);
   row1 = _mm_shufflelo_epi16(left_col, 0x55);
   row2 = _mm_shufflelo_epi16(left_col, 0xaa);
   row3 = _mm_shufflelo_epi16(left_col, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
 }
 
 void aom_h_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
@@ -1023,13 +1023,13 @@ void aom_h_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   __m128i row1 = _mm_shufflelo_epi16(left_col_low, 0x55);
   __m128i row2 = _mm_shufflelo_epi16(left_col_low, 0xaa);
   __m128i row3 = _mm_shufflelo_epi16(left_col_low, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
   dst += stride;
 
   left_col_low = _mm_unpackhi_epi64(left_col_low, left_col_low);
@@ -1037,26 +1037,26 @@ void aom_h_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   row1 = _mm_shufflelo_epi16(left_col_low, 0x55);
   row2 = _mm_shufflelo_epi16(left_col_low, 0xaa);
   row3 = _mm_shufflelo_epi16(left_col_low, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
   dst += stride;
 
   row0 = _mm_shufflelo_epi16(left_col_high, 0);
   row1 = _mm_shufflelo_epi16(left_col_high, 0x55);
   row2 = _mm_shufflelo_epi16(left_col_high, 0xaa);
   row3 = _mm_shufflelo_epi16(left_col_high, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
   dst += stride;
 
   left_col_high = _mm_unpackhi_epi64(left_col_high, left_col_high);
@@ -1064,13 +1064,13 @@ void aom_h_predictor_4x16_sse2(uint8_t *dst, ptrdiff_t stride,
   row1 = _mm_shufflelo_epi16(left_col_high, 0x55);
   row2 = _mm_shufflelo_epi16(left_col_high, 0xaa);
   row3 = _mm_shufflelo_epi16(left_col_high, 0xff);
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row0);
+  *(int *)dst = _mm_cvtsi128_si32(row0);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row1);
+  *(int *)dst = _mm_cvtsi128_si32(row1);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row2);
+  *(int *)dst = _mm_cvtsi128_si32(row2);
   dst += stride;
-  *(uint32_t *)dst = _mm_cvtsi128_si32(row3);
+  *(int *)dst = _mm_cvtsi128_si32(row3);
 }
 
 void aom_h_predictor_8x4_sse2(uint8_t *dst, ptrdiff_t stride,
